input-buffer: append gestures in a primary method
Previously we have appended gestures from the :after method. That lead to a
situation, where drei gadgets which are input streams (most notabli text-field
and text-editor panes) first handled the gesture by adding a character to the
gadget, and then append said gesture to the input-buffer, which was then
dequeued by the input-editing stream from the interactor and modified the
input of the interactor. In other words gestures were distributed twice.